refactor(core,inference): move FLUX bundled-VAE probe to mold-core
